Added heart buttons
I've created a heart button that appears to fill on hover. To do this I have used the heart outline and full heart icons from fontawesome and animated the full heart to go from 0 to 1 opacity on hover, therefore giving the illusion that the heart is filling. I'm not sure this is the best solution and if I find a better way I will update them