Added first draft of the tick transition
I've added the tick transition but its not fully finished. The problem I'm facing is that the card text does not react to the tick when it slides on to the menu card, causing it to become hidden. I know this is because the tick is positioned absolutely, but I'm finding it difficult to find another way to implement it with different positioning.